requirements.txt
Django==1.8.13
psycopg2
bcrypt
bitstring
dpkt
celery
pytz
python-dateutil
pycrypto
redis
pillow==3.1.1
django-environ
django-celery-results
django-suit
django-redis-cache==1.6.4
django-autocomplete-light==2.3.3
django-cors-headers
django-crispy-forms==1.6.0
django-markdown-deux
django-recaptcha
django-colorfield
django-timezone-field
django_compressor
djangorestframework>=3.5.3
djangorestframework-jwt
django-filter>=1.0.1
Django-Select2
drf_ujson
sympy>=1.0.0
